create post modified

+++ app/controllers/likes_controller.rb
@@ -3,29 +3,29 @@ class LikesController < ApplicationController
   before_action :authenticate_user!
 
 	def toggle_like
-		@post = Post.find(params[:post_id])
-		like = Like.where(user: current_user, post: @post).first
-		dislike = Dislike.where(user: current_user, post: @post).first
-		if like
-			like.destroy!
-			@is_liked = false
-		else
-			Like.create(user: current_user, post: @post)
-			
-			if dislike
-				dislike.destroy!
-				@is_disliked = false
-				puts "in dislike"
+    if params[:type]=="post"
+    		@post = Post.find(params[:post_id])
+    		like = Like.where(user: current_user, post: @post).first
+    		dislike = Dislike.where(user: current_user, post: @post).first
+    		if like
+    			like.destroy!
+    			@is_liked = false
+    		else
+    			Like.create(user: current_user, post: @post)
 
-			end
-			@is_liked = true
-		end
+    			if dislike
+    				dislike.destroy!
+    				@is_disliked = false
+    				puts "in dislike"
 
-		respond_to do |format|
-			format.js { }
+    			end
+    			@is_liked = true
+    		end
 
-		end
+    		respond_to do |format|
+    			format.js { }
 
-
-	end
+    		end
+    	end
+    end
 end
